backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20  # required by FastAPI for file upload / form data endpoints

# Validation
pydantic[email]==2.10.3

# PostgreSQL
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
# psycopg2-binary: needed by Alembic for sync migrations.
# The binary wheel is pre-built and works on all platforms including Windows.
psycopg2-binary==2.9.10

# pgvector (scaffolded now; populated in Phase 3)
pgvector==0.3.6

# Neo4j
neo4j==5.26.0
neomodel==5.4.1

# Task queue (Celery + Redis broker)
celery[redis]==5.4.0
redis==5.2.1

# Object storage (S3-compatible: MinIO locally, Cloudflare R2 in production)
aioboto3==13.2.0

# Music processing
music21==9.3.0
lxml==5.3.0
pyyaml==6.0.2

# Score rendering (server-side: incipit generation, fragment preview)
verovio==6.1.0

# Auth / JWT (Supabase JWT validation)
python-jose[cryptography]==3.3.0
httpx==0.28.1

# Environment variable loading (.env files in dev; no-op in production)
python-dotenv==1.1.0

